requirements.txt
Flask
boto3
pymysql
python-dotenv
cryptography